Move SPIRVLowerConstExprPass to header file
So that it could be used in downstream.
